🛠  MULE-IOS [MODIFIED] ios emulator list function moved to ios folder
