+++ src/index.js
@@ -1,53 +1,11 @@
 const inquirer = require("inquirer");
 const { exec } = require("child_process");
 const { exit } = require("process");
-const { iosRuntimeList } = require("./ios");
+const { iosRuntimeList, iosEmulatorList } = require("./ios");
 const { SCRIPT_PREFIX, OS_TYPE_Q } = require("./constants");
 
 const pickOsType = () => {
   return OS_TYPE_Q;
-};
-
-const iosEmulatorList = (runtimeKey) => {
-  return new Promise((resolve, reject) => {
-    exec(`sh ${SCRIPT_PREFIX}emu_list_ios.sh`, (err, stdout, stderr) => {
-      if (err) {
-        console.error(err);
-        reject(err);
-        exit;
-      } else {
-        if (stdout) {
-          const { devices } = JSON.parse(stdout);
-
-          const bootedDevices = devices[runtimeKey].filter(
-            (v) => v.state !== "Shutdown"
-          );
-          const offDevices = devices[runtimeKey]
-            .filter((v) => v.state !== "Booted")
-            .sort((a, b) => (a.name < b.name ? 1 : -1));
-
-          //   console.log("Booted Devices", bootedDevices);
-          //   console.log("Off Devices", offDevices);
-          const sortedList = [...bootedDevices, ...offDevices];
-
-          resolve({
-            type: "list",
-            name: "ios_device",
-            message: "Select an IOS Device",
-            choices: sortedList.map((device) => ({
-              key: device.udid,
-              name: `${device.name} -> (${device.state})`,
-              value: JSON.stringify({
-                deviceName: device.name,
-                deviceUDID: device.udid,
-                deviceState: device.state,
-              }),
-            })),
-          });
-        }
-      }
-    });
-  });
 };
 
 const androidEmulatorList = () => {
